Moved config stuff to corresponding file and refined some tests
